requirements/base.txt
arrow==1.2.3
asgiref==3.6.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
Django==3.2.18
django-constance==2.7.0
django-debug-toolbar==4.0.0
django-extensions==3.0.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-picklefield==3.0.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2023.3
urllib3==1.26.15
